requirements/dev.txt
-r common.txt
ipython==7.13.0
pytest==5.4.2
pytest-django==3.9.0